Move simpleStream snippets to WearableStreams
Split the simple stream experiment programs out of WearableExample
and into WearableStreams. Keep the functions that are exported for
other programs (Main, Criterion).
